requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
scipy>=1.7.0
pyyaml>=6.0
pathlib>=1.0.1 